Move postprocessing and assembly related things from StructuralHarmonicSolver
